+++ actions.js
@@ -352,7 +352,7 @@ export const actions = {
                     global.evolution['final'] = 100;
                     global.evolution['sentience'] = { count: 0 };
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         if (global.evolution['chitin']){
                             global.evolution['sporgar'] = { count: 0 };
                             global.evolution['shroomi'] = { count: 0 };
@@ -398,7 +398,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['mantis'] = { count: 0 };
                         global.evolution['scorpid'] = { count: 0 };
                         global.evolution['antid'] = { count: 0 };
@@ -478,7 +478,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['human'] = { count: 0 };
                         global.evolution['orc'] = { count: 0 };
                         global.evolution['elven'] = { count: 0 };
@@ -520,7 +520,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['troll'] = { count: 0 };
                         global.evolution['orge'] = { count: 0 };
                         global.evolution['cyclops'] = { count: 0 };
@@ -562,7 +562,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['kobold'] = { count: 0 };
                         global.evolution['goblin'] = { count: 0 };
                         global.evolution['gnome'] = { count: 0 };
@@ -604,7 +604,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['cath'] = { count: 0 };
                         global.evolution['wolven'] = { count: 0 };
                         global.evolution['centaur'] = { count: 0 };
@@ -644,7 +644,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['balorg'] = { count: 0 };
                         global.evolution['imp'] = { count: 0 };
                         addAction('evolution','balorg');
@@ -680,7 +680,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['sharkin'] = { count: 0 };
                         global.evolution['octigoran'] = { count: 0 };
                         addAction('evolution','sharkin');
@@ -742,7 +742,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['arraak'] = { count: 0 };
                         global.evolution['pterodacti'] = { count: 0 };
                         global.evolution['dracnid'] = { count: 0 };
@@ -776,7 +776,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['tortoisan'] = { count: 0 };
                         global.evolution['gecko'] = { count: 0 };
                         global.evolution['slitheryn'] = { count: 0 };
@@ -9421,6 +9421,10 @@ function bioseed(){
     unlockAchieve(`seeder`);
     let new_biome = unlockAchieve(`biome_${biome}`);
     let new_genus = unlockAchieve(`genus_${genus}`);
+    let probes = global.starDock.probes.count + 1;
+    if (global.stats.achieve['explorer']){
+        probes += global.stats.achieve['explorer'];
+    }
     global['race'] = { 
         species : 'protoplasm', 
         gods: god,
@@ -9428,7 +9432,7 @@ function bioseed(){
         Plasmid: { count: plasmid },
         Phage: { count: phage },
         seeded: true,
-        probes: global.starDock.probes.count + 1,
+        probes: probes,
         seed: Math.floor(Math.seededRandom(10000)),
     };
     global.city = {