+++ changelog.md
@@ -1,5 +1,7 @@
 #### 0.?.?? - 7/??/2019 ####
 Slow and Hyper traits can now combine<br>
+Added new achievements: Creator & Explorer<br>
+Added perks for Mass Extinction, Creator, and Explorer<br>
 
 #### 0.4.35 - 7/25/2019 ####
 Centaur can now pick Fanaticism<br>