+++ changelog.md
@@ -1,4 +1,4 @@
-#### 0.?.?? - 7/??/2019 ####
+#### 0.4.36 - 7/27/2019 ####
 Slow and Hyper traits can now combine<br>
 Added new achievements: Creator & Explorer<br>
 Added perks for Mass Extinction, Creator, and Explorer<br>